test(web-serial-data-access): unify pipeline spec files
Rename the pipeline spec to match the service name and merge queue-focused assertions into the same spec file to keep a 1:1 service-spec mapping.
